% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsparser-package.R
\docType{package}
\name{bigsparser-package}
\alias{bigsparser}
\alias{bigsparser-package}
\title{bigsparser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
