% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFBM.R
\docType{class}
\name{SFBM_compact-class}
\alias{SFBM_compact-class}
\alias{SFBM_compact_RC}
\title{Class SFBM_compact}
\description{
A reference class for storing and accessing sparse matrix-like data stored
in files on disk, in a compact format (when non-zero values in columns are
contiguous).
}
\details{
It inherits the fields and methods from class \link[=SFBM-class]{SFBM}.
}

