% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsparser-package.R
\docType{package}
\name{bigsparser-package}
\alias{bigsparser}
\alias{bigsparser-package}
\title{bigsparser: Sparse Matrix Format With Data On Disk}
\description{
Provides a sparse matrix format with data stored on disk, to be
    used in both R and C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/bigsparser}
  \item Report bugs at \url{https://github.com/privefl/bigsparser/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
