% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFBM.R
\docType{class}
\name{SFBM-class}
\alias{SFBM-class}
\alias{SFBM_RC}
\alias{as_SFBM}
\title{Class SFBM}
\usage{
as_SFBM(spmat, backingfile = tempfile())
}
\arguments{
\item{spmat}{A dgCMatrix (non-symmetric sparse matrix of type 'double').}

\item{backingfile}{Path to file where to store data. Extension \code{.sbk} is
automatically added.}
}
\value{
The new \link[=SFBM-class]{SFBM}.
}
\description{
A reference class for storing and accessing sparse matrix-like data stored
in files on disk.

Convert a dgCMatrix to an SFBM.
}
\details{
An object of class SFBM has many fields:
\itemize{
\item \verb{$address}: address of the external pointer containing the underlying
C++ object to be used as a \verb{XPtr<SFBM>} in C++ code
\item \verb{$extptr}: (internal) use \verb{$address} instead
\item \verb{$nrow}: number of rows
\item \verb{$ncol}: number of columns
\item \verb{$nval}: number of non-zero values
\item \verb{$p}: vector of column positions
\item \verb{$backingfile} or \verb{$sbk}: File with extension 'sbk' that stores the
data of the SFBM
\item \verb{$rds}: 'rds' file (that may not exist) corresponding to the 'sbk' file
\item \verb{$is_saved}: whether this object is stored in \verb{$rds}?
}

And some methods:
\itemize{
\item \verb{$save()}: Save the SFBM object in \verb{$rds}. Returns the SFBM.
}
}

\examples{
spmat <- Matrix::rsparsematrix(1000, 1000, 0.01)
class(spmat)
(X <- as_SFBM(spmat))

}
