% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFBM.R
\name{dim,SFBM-method}
\alias{dim,SFBM-method}
\alias{length,SFBM-method}
\title{Dimension and type methods for class \code{SFBM}.}
\usage{
\S4method{dim}{SFBM}(x)

\S4method{length}{SFBM}(x)
}
\arguments{
\item{x}{An object of class \link[=SFBM-class]{SFBM}.}
}
\description{
Dimension and type methods for class \code{SFBM}.
}
